fix activity update for community solid server
