+++ test/components/Users/ManageUsers.test.jsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import React from 'react';
 import { expect, it, describe } from 'vitest';
-import ManageUsers  from '../../../../src/components/Users/ManageUsers';
+import ManageUsers from '../../../src/components/Users/ManageUsers';
 
 describe('When page loads', () => {
   it('renders correctly', () => {
